% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_arima_utils.R
\name{calibrar.arima}
\alias{calibrar.arima}
\title{Best parameters arima model}
\usage{
calibrar.arima(train, test, period, ar = 0:2, es = 0:1)
}
\arguments{
\item{train}{a ts object (train of a time series).}

\item{test}{a ts object (test of a time series).}

\item{period}{value indicate the period to use.}

\item{ar}{vector of values to test p, d, q of arima model.}

\item{es}{vector of values to test P, D, Q of arima model.}
}
\value{
arima model
}
\description{
Best parameters arima model
}
\examples{
calibrar.arima(AirPassengers[1:132], AirPassengers[133:144], 12, 0:1)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
