% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_holtwinters_utils.R
\name{calibrar.HW}
\alias{calibrar.HW}
\title{Best parameters HoltWinters model}
\usage{
calibrar.HW(train, test, paso = 0.1)
}
\arguments{
\item{train}{a ts object (train of a time series).}

\item{test}{a ts object (test of a time series).}

\item{paso}{indicates by value to test alpha, beta and gamma.}
}
\value{
HoltWinters model
}
\description{
Best parameters HoltWinters model
}
\examples{
calibrar.HW(window(AirPassengers, end = c(1959, 12)), window(AirPassengers, start = 1960), 0.5)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
