% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_periodograma_utils_.R
\name{df_periods}
\alias{df_periods}
\title{Periodogram Data.frame}
\usage{
df_periods(x)
}
\arguments{
\item{x}{a ts object.}
}
\value{
data.frame
}
\description{
Periodogram Data.frame
}
\examples{
df_periods(AirPassengers)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
