% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_arima_utils.R
\name{e_acf}
\alias{e_acf}
\title{Best parameters arima model}
\usage{
e_acf(x)
}
\arguments{
\item{x}{a ts object.}
}
\value{
echarts4r plot
}
\description{
Best parameters arima model
}
\examples{
e_acf(AirPassengers)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
