% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t_c_utils.R
\name{e_tc}
\alias{e_tc}
\title{Tendencia y Estacionalidad}
\usage{
e_tc(x, d = NULL, noms = c("Time Series", "Trend", "Cyclicality"))
}
\arguments{
\item{x}{a ts object.}

\item{d}{a vector of dates to use on axis x (Optional).}

\item{noms}{a vector of 3 to indicate the names to use on legend.}
}
\value{
data.frame
}
\description{
Tendencia y Estacionalidad
}
\examples{
e_tc(AirPassengers)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
