% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasteR.R
\name{estimateModels}
\alias{estimateModels}
\title{Estimate statistical models on data based on a given list of relationships}
\usage{
estimateModels(data, formulas)
}
\arguments{
\item{data}{A dataframe with some blanks in endogenous variables, and a period column}

\item{formulas}{A list of formulas (e.g., list(x = price ~ weather + month, y = demand ~ price + income))}
}
\value{
A formula
}
\description{
\code{estimateStatisticalModels} returns a list of statistical models.
}
