% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasteR.R
\name{forecast}
\alias{forecast}
\title{Forecast missing values based on a model (set of equations).}
\usage{
forecast(data, model, periods, periodColumn, transformation = NULL,
  lowerBounds, upperBounds, solver)
}
\arguments{
\item{data}{A dataframe with some blanks in endogenous variables, and a period column}

\item{model}{A list of equations specifying the relationships among endogenous variables. The names of equations specify endogenous variables that they solve.}

\item{periods}{For which periods should the simulation be run?}

\item{periodColumn}{The name of the column with the period values}

\item{transformation}{A function specifying how the data need to be transformed post each simulation (e.g., define lagged variables)}

\item{lowerBounds}{An optional list of lower bounds for endogenous variables}

\item{upperBounds}{An optional list of upper bounds for endogenous variables}

\item{solver}{A function that can solve a system of equations (accepts model, endogenous, exogenous and return list with "result" as a list of the same structure as endogenous)}
}
\value{
A data frame
}
\description{
\code{forecast} returns the dataframe after the the missing values have been filled using the provided model.
}
