% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasteR.R
\docType{package}
\name{forecasteR}
\alias{forecasteR}
\title{Time Series Forecast System}
\description{
Performs a time series forecast.
}
\details{
\tabular{ll}{
Package: \tab forecasteR\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2021-06-04\cr
License: \tab GPL (>=2)\cr
}
}
\author{
Oldemar Rodriguez Rojas \cr
Maintainer: Oldemar Rodriguez Rojas <oldemar.rodriguez@ucr.ac.cr>
}
\keyword{package}
