% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_carga_datos_utils.R
\name{get_start}
\alias{get_start}
\title{Get ts start of a time series}
\usage{
get_start(ini, tipo_f, patron)
}
\arguments{
\item{ini}{a Date object.}

\item{tipo_f}{type of the time series ('year', 'month', ..., 'seconds').}

\item{patron}{frequency of time series.}
}
\value{
numeric vector of lenght 2
}
\description{
Get ts start of a time series
}
\examples{
get_start(as.Date("2021-06-30"), 'days', 365)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
