% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_utils.R
\name{grafico.errores}
\alias{grafico.errores}
\title{Error plot for all predictions}
\usage{
grafico.errores(errores)
}
\arguments{
\item{errores}{a data.frame with errors of a time series.}
}
\value{
data.frame
}
\description{
Error plot for all predictions
}
\examples{
model <- arima(window(AirPassengers, end = c(1959, 12)))
pred  <- predict(model, 12)
e <- tabla.errores(list(pred$pred), window(AirPassengers, start = 1960))
grafico.errores(e)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
