% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_periodograma.R
\name{mod_periodograma_server}
\alias{mod_periodograma_server}
\title{periodograma Server Function}
\usage{
mod_periodograma_server(input, output, session, updateData)
}
\description{
periodograma Server Function
}
