% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_periodograma_utils_.R
\name{plot_periods}
\alias{plot_periods}
\title{Periodogram Plot}
\usage{
plot_periods(x, p = NULL)
}
\arguments{
\item{x}{a ts object.}

\item{p}{which important period to plot.}
}
\value{
echarts4r plot
}
\description{
Periodogram Plot
}
\examples{
plot_periods(AirPassengers)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
