% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_utils.R
\name{PTFA}
\alias{PTFA}
\title{Percentage of Total Failures Up}
\usage{
PTFA(Pred, Real)
}
\arguments{
\item{Pred}{a ts object (prediction).}

\item{Real}{a ts object (real).}
}
\value{
numeric
}
\description{
Percentage of Total Failures Up
}
\examples{
model <- arima(window(AirPassengers, end = c(1959, 12)))
pred  <- predict(model, 12)
PTFA(pred$pred, window(AirPassengers, start = 1960))

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
