% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_utils.R
\name{RE}
\alias{RE}
\title{Relative Error}
\usage{
RE(Pred, Real)
}
\arguments{
\item{Pred}{a ts object (prediction).}

\item{Real}{a ts object (real).}
}
\value{
numeric
}
\description{
Relative Error
}
\examples{
model <- arima(window(AirPassengers, end = c(1959, 12)))
pred  <- predict(model, 12)
RE(pred$pred, window(AirPassengers, start = 1960))

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
