% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_carga_datos_utils.R
\name{smoothing}
\alias{smoothing}
\title{Apply rolling to a numeric vector.}
\usage{
smoothing(v, n)
}
\arguments{
\item{v}{a numeric vector.}

\item{n}{integer value specifying the window width.}
}
\value{
numeric vector
}
\description{
Apply rolling to a numeric vector.
}
\examples{
smoothing(AirPassengers, 5)

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
