% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_utils.R
\name{tabla.errores}
\alias{tabla.errores}
\title{Error table for all predictions}
\usage{
tabla.errores(Preds, Real, nombres = NULL)
}
\arguments{
\item{Preds}{a list of ts objects (prediction).}

\item{Real}{a ts object (real).}

\item{nombres}{names for the data.frame (optional).}
}
\value{
data.frame
}
\description{
Error table for all predictions
}
\examples{
model <- arima(window(AirPassengers, end = c(1959, 12)))
pred  <- predict(model, 12)
tabla.errores(list(pred$pred), window(AirPassengers, start = 1960))

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
