% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_carga_datos_utils.R
\name{text_toDate}
\alias{text_toDate}
\title{Convert character to dates}
\usage{
text_toDate(f)
}
\arguments{
\item{f}{a vector of character.}
}
\value{
list
}
\description{
Convert character to dates
}
\examples{
text_toDate(c("2023 january 27", "2023 january 28"))

}
\author{
Diego Jimenez <diego.jimenez@promidat.com>
}
